+++ include/kernel/tty.h
@@ -68,8 +68,8 @@ void tty_init();
 void tty_fs_init();
 void tty_move_cursor(tty_t* tty, int x, int y);
 void tty_handle_input(input_event_t input);
-uint32_t tty_write(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer);
-uint32_t tty_read(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer);
+uint64_t tty_write(fs_node_t *node, uint64_t offset, uint64_t size, uint8_t *buffer);
+uint64_t tty_read(fs_node_t *node, uint64_t offset, uint64_t size, uint8_t *buffer);
 void tty_set_cursor_visibility(tty_t* tty, int visible);
 void tty_clear_screen(tty_t* tty);
